Fix the http package server
The test server just dies unhelpfully when something goes wrong,
additionally it appears that the file length being returned by
Files.size in JDK17 aren't what we expect, and can't be used a response
length.

To make the server a little more resilient and helpful, it now returns a
405 Method Not Allowed if it gets anything except a "GET", rather than
throwing and dying. Also doesn't set the response length any more, as
that was causing the server to throw while returning a package, which
then closed the on-going connection early and killed the server for
future tests.

The mock servers is also much more aggressive at starting up, if it can't
bind to the needed port, it'll keep trying for 20 seconds, with 1 second
between retries. That way if there are multiple tests trying to create a
server in parallel, this will effectively serialize the test runs, but
prevent one set of tests failing because it was relying an server
created elsewhere that then shutdown during test execution.