Fix name resolution in `typealias` with constraint
The body of a `typealias` gets inlined into wherever the `typealias`
is used. This fixes a bug where the name resolution would resolve an
additional type constraint on the next scope of `this`.

Lack of a frame scope
---------------------

`typealias` doesn't currently have a frame scope, which is why we have to assign
the `owner` and the `receiver` here. If it did, we could read both of these
at the point where we're trying to resolve the `typealias` in question.