Overhaul typechecks for Mapping/Listing
This changes how the language performs typechecks for mappings and
listings.

Currently, Pkl will shallow-force any Mapping and Listing to check it
the type parameter (e.g. Listing<Person> means each element is checked
to be an instance of Person).

This changes the language to check each member's type when the member
is accessed.

Note: List/Map/Set are still eager checks, because these types are not
VmObject and not lazy in the first place.

Note 2: Unions of Listings or Mappings are still eager (we must force
the listing to know if it satisfies Listing<A>|Listing<B>).