Refactor ResourceReader to support multiple schemes
This is preparatory work for [SPICE-0009](https://github.com/apple/pkl-evolution/pull/10).
It is being contributed in a separate pull request to ease review.

This is required to allow deferred launch of external reader processes without requiring up-front configuration of scheme->process mappings.
More details on why this is required here: https://github.com/apple/pkl-evolution/pull/10#discussion_r1724145677

One concern here (which applies before this change as well) is that the semantics of module key factories and resource readers is reversed.
For module key factories, the first factory that answers for a URI scheme "wins", while for resource readers, the _last_ reader that answers for a URI scheme "wins".
This PR preserves that behavior, but it may be worth reconsidering this design at this time.
This behavior can be observed in practice in `ResourceReadersEvaluatorTest.\`module path\`` where `ResourceReaders.modulePath` is added after the pre-configured `ResourceReaders.classPath` and takes precedence.