Fix double unary minus
If AstBuilder is processing UnaryMinus -> UnaryMinus -> IntLiteral, the "inner" UnaryMinus helpfully defers inversion to the IntLiteral and returns it directly.
The "_outer_" UnaryMinus, none the wiser, then does the exact same thing.
There needs to be a way to force only the inner-most UnaryMinus to delegate to the literal node, which seems tricky in context.

This is my (relatively naive) attempt at that, but this feels messy.
I think there's also a possible overflow edge case around minInt.

More than happy to take suggestions on a better way to handle this!

Closes #653