Vendor msgpack-java
This adds msgpack-java's sources to pkl-core, which allows us to exclude this as a dependency.

This follows a pattern that we have with a few other libraries (JSON, YAML).

Note: package `org.pkl.core.util.msgpack` is considered private, but yet is used in `pkl-server`.
In the future, we plan on moving all messagepack serializing/deserializing to pkl-core.