Use `BuildInfo.GraalVm` parameter to construct the output files
Make sure that `graal_isolate.h` and `graal_isolate_dynamic.h` are
tracked as output files.

Extension adjustment
--------------------

`native-image` will add an extension to the filename provided `-o`. We
need to make sure that we track this correctly for each OS, and remove
the extension when we provide it to `native-image`.