Treat opaque file URIs as errors
Opaque file URIs are URIs whose scheme-specific part does not start with
`/`.
For example, `file:foo/bar.txt` is an opaque URIs.

Currently, this has the unintentional behavior of: look for file
`foo/bar.txt` from the process working directory.
These are effectively dynamics imports; from a single import, we can't
statically analyze what it resolves as.

According to RFC-8089, File URIs must have paths that start with `/`.
So, opaque file URIs are actually not valid.

This changes Pkl's behavior so that these URIs are treated as errors.